Lathe logic fix & Main menu rework
Lathe logic fixed so that select program doesn't require an item in lathe to work.
Main menu rework with buttons moved and Test mode toggle switched out to an own button to start in test mode.

Rework idea is to handle future game updates with different machines added to the base game,